♻️ Move modal & elements file
